リポジトリではなくクエリサービスを利用するように修正
- ユースケースの名前が長いなと思ったので、修正
- team-repository → get-all-member-query-service
