+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/test/stub/infra/db/query-service/get-all-member-query-service.ts
