+++ 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/usecase/query-service-interface/get-all-member-query-service.ts
@@ -1,0 +1,11 @@
+import { Member } from "domain/team/entity/member";
+import { Team } from "domain/team/entity/team";
+
+export interface ITeamStructure {
+  teamList: Team[];
+  independentMemberList?: Member[];
+}
+
+export interface IGetAllMemberQueryService {
+  execute(): Promise<ITeamStructure>;
+}
