--- 04_設計/04_プラハチャレンジをDDDで実装してみる/02_APIサーバ/src/usecase/team/get-all-team-and-member.ts
@@ -1,15 +1,0 @@
-import { ITeamRepository, ITeamStructure } from "domain/team/team-repository";
-
-export class GetAllTeamAndMember {
-  private readonly teamRepository: ITeamRepository;
-
-  constructor(teamRepository: ITeamRepository) {
-    this.teamRepository = teamRepository;
-  }
-
-  public async execute(): Promise<ITeamStructure> {
-    const result = await this.teamRepository.getAll();
-
-    return result;
-  }
-}
